Move badge update to tray
