+++ electron/js/lib/download.js
@@ -54,4 +54,3 @@ module.exports = function(fileName, bytes) {
     });
   });
 };
-
