+++ electron/locale/strings-en.js
@@ -56,4 +56,6 @@ string.restartLocale = 'Please restart the app for this setting to take effect.'
 string.restartLater = 'Later';
 string.restartNow = 'Restart Now';
 
+string.unreadMessages = 'Unread message';
+
 module.exports = string;
