+++ electron/locale/strings.js
@@ -53,3 +53,5 @@ string.restartNeeded = 'Restart Needed';
 string.restartLocale = 'Please restart the app for this setting to take effect.';
 string.restartLater = 'Later';
 string.restartNow = 'Restart Now';
+
+string.unreadMessages = 'Unread message';
