+++ electron/main.js
@@ -144,8 +144,6 @@ function showMainWindow() {
     },
   });
 
-  let unread_count = 0;
-
   if (init.restore('fullscreen', false)) {
     main.setFullScreen(true);
   } else {
@@ -203,14 +201,12 @@ function showMainWindow() {
     }
   });
 
-  main.on('focus', function(){
+  main.on('focus', function() {
     main.flashFrame(false);
   });
 
   main.on('page-title-updated', function() {
-    util.updateBadge(main, unread_count).then(function(new_unread_count) {
-      unread_count = new_unread_count;
-    });
+    tray.updateBadgeIcon(main);
   });
 
   main.on('close', function(event) {
